+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+ - better console log filtering when not in `--verbose` mode
 
 
 ## [3.0.0] - 2026-06-22